+++ cypress-shared/integration/fedex-shipping/2.10-validate-surcharge-flat-rate.spec.js
@@ -1,9 +1,13 @@
 /* eslint-disable jest/valid-expect */
 /* eslint-disable jest/valid-expect-in-promise */
+/* eslint-disable jest/expect-expect */
 import { updateRetry, loginViaCookies } from '../../support/common/support'
 import { appSetting } from '../../support/fedex-shipping/outputvalidation'
 import { data } from '../../fixtures/fedex-shipping-fixtures/shippingRatePayload.json'
-import { updateSLASettings } from '../../support/fedex-shipping/common.js'
+import {
+  filterByShippingMethod,
+  updateSLASettings,
+} from '../../support/fedex-shipping/common.js'
 import {
   loadCalculateShippingAPI,
   validateCalculateShipping,
@@ -25,11 +29,9 @@ describe('Modify SLA - Validate Surcharge Flat Rate in checkout', () => {
   it(`${prefix} - Verify shipping price`, updateRetry(3), () => {
     loadCalculateShippingAPI(data).then((response) => {
       validateCalculateShipping(response)
-      const filtershippingMethod = response.body.filter(
-        (b) => b.shippingMethod === sla.FirstOvernight
-      )
+      const shippingMethods = filterByShippingMethod(response, sla)
 
-      amount = filtershippingMethod[0].price
+      amount = shippingMethods[0].price
     })
   })
 
@@ -40,13 +42,11 @@ describe('Modify SLA - Validate Surcharge Flat Rate in checkout', () => {
   it(`${prefix} - Validate Surcharge Flat Rate Changes`, updateRetry(3), () => {
     loadCalculateShippingAPI(data).then((response) => {
       validateCalculateShipping(response)
-      const filtershippingMethod = response.body.filter(
-        (b) => b.shippingMethod === sla.FirstOvernight
-      )
+      const shippingMethods = filterByShippingMethod(response, sla)
 
       const calculateFlatRate = amount + surchargeFlatRate
 
-      expect(filtershippingMethod[0].price).to.equal(calculateFlatRate)
+      expect(shippingMethods[0].price).to.equal(calculateFlatRate)
     })
   })
 })