+++ cypress-shared/integration/fedex-shipping/2.3-single-product-scenarios.spec.js
@@ -1,3 +1,4 @@
+/* eslint-disable jest/expect-expect */
 /* eslint-disable jest/valid-expect */
 /* eslint-disable jest/valid-expect-in-promise */
 import { loginViaCookies, updateRetry } from '../../support/common/support.js'
@@ -11,15 +12,15 @@ import {
   validateCalculateShipping,
 } from '../../support/fedex-shipping/api_testcase.js'
 import {
-  graphql,
   verifyInventoryIsUnlimitedForFedexWareHouse,
   validateInventory,
 } from '../../support/fedex-shipping/graphql_testcase.js'
 import { INVENTORY_GRAPHQL_APP } from '../../support/fedex-shipping/graphql_apps.js'
 import sla from '../../support/fedex-shipping/sla.js'
+import { graphql } from '../../support/common/graphql_utils'
 
 const { prefix } = singleProduct
-let amount = ''
+let shippingMethod = {}
 
 describe(`${prefix} Scenarios`, () => {
   loginViaCookies()
@@ -39,10 +40,12 @@ describe(`${prefix} Scenarios`, () => {
     loadCalculateShippingAPI(data).then((response) => {
       validateCalculateShipping(response)
       const filtershippingMethod = response.body.filter(
-        (b) => b.shippingMethod === sla.FirstOvernight
+        (b) =>
+          b.shippingMethod === sla.FirstOvernight ||
+          b.shippingMethod === sla.StandardOvernight
       )
 
-      amount = filtershippingMethod[0].price
+      shippingMethod = filtershippingMethod[0]
     })
   })
 
@@ -54,10 +57,15 @@ describe(`${prefix} Scenarios`, () => {
       loadCalculateShippingAPI(data).then((response) => {
         validateCalculateShipping(response)
         const filtershippingMethod = response.body.filter(
-          (b) => b.shippingMethod === sla.FirstOvernight
+          (b) =>
+            b.shippingMethod === sla.FirstOvernight ||
+            b.shippingMethod === sla.StandardOvernight
         )
 
-        expect(filtershippingMethod[0].price).to.equal(amount * 2)
+        expect(shippingMethod.shippingMethod).to.equal(
+          filtershippingMethod[0].shippingMethod
+        )
+        expect(filtershippingMethod[0].price).to.equal(shippingMethod.price * 2)
       })
     }
   )