+++ cypress-shared/integration/quick-order/1.3.1-upload_csv.spec.js
@@ -3,19 +3,15 @@ import {
   preserveCookie,
   updateRetry,
 } from '../../support/common/support.js'
-import {
-  quickOrderByXLS,
-  quickOrderByXLSNegativeTestCase2,
-} from '../../support/quick-order/testcase.js'
+import { quickOrderByXLS } from '../../support/quick-order/testcase.js'
 
-describe('Quickorder - SkuCode Quantity testcase', () => {
+describe('Quickorder - Upload CSV testcase', () => {
   loginViaCookies()
 
-  quickOrderByXLSNegativeTestCase2(false)
   quickOrderByXLS(false)
 
   it(
-    `In SkuCode Quantity testcase - Verify quantity by price`,
+    `In Upload CSV testcase - Verify quantity by price`,
     updateRetry(4),
     () => {
       cy.get('.quantity.badge').first().should('have.text', '2')