+++ cypress-shared/support/common/testcase.js
@@ -221,26 +221,6 @@ export function startPaymentE2ETests() {
   // then ensure, you are having affiliationId, apiKey and apiToken in .VTEX_QE.json
   it(`Start ${prefix} E2E tests with this workspace ${WORKSPACE}`, () => {
     verifyEnvs(true)
-    cy.request({
-      url: affiliationAPI(affiliationId),
-      headers: VTEX_AUTH_HEADER(apiKey, apiToken),
-      ...FAIL_ON_STATUS_CODE,
-    }).then((response) => {
-      const workspaceIndex = response.body.configuration.findIndex(
-        (obj) => obj.name === 'workspace'
-      )
-
-      const workspaceInGatewayAffialitions =
-        response.body.configuration[workspaceIndex].value
-
-      if (!['', WORKSPACE].includes(workspaceInGatewayAffialitions)) {
-        throw new Error(
-          `Another test is running with workspace ${workspaceInGatewayAffialitions}. Please, try again later.\n`
-        )
-      } else {
-        cy.log('Starting E2E configuration')
-      }
-    })
   })
 }