+++ cypress-shared/support/fedex-shipping/outputvalidation.js
@@ -13,6 +13,19 @@ export default {
     unitWeight: 'KG',
     unitDimension: 'CM',
     packingAccessKey: '',
+    defaultDeliveryEstimateInDays: '0',
+  },
+  packingOptimizationSettings: {
+    accessKey: 'fedexdev',
+    containerList: [
+      {
+        id: 1,
+        length: 20,
+        width: 20,
+        height: 20,
+        description: 'cypress test',
+      },
+    ],
   },
   smartPackingAccessKey: 'fedexdev',
   docks: {
@@ -24,6 +37,16 @@ export default {
     fedexBrazilDock: {
       id: 'S0001-E0001-Dock',
       name: 'Fedex Italy Dock',
+      shippingRatesProviders: ['vtexus.fedex-shipping'],
+    },
+    fedexIndiaDock: {
+      id: 'IND00112233',
+      name: 'Fedex India Dock',
+      shippingRatesProviders: ['vtexus.fedex-shipping'],
+    },
+    fedexKuwaitDock: {
+      id: 'KUWA0112233',
+      name: 'Fedex Kuwait Dock',
       shippingRatesProviders: ['vtexus.fedex-shipping'],
     },
   },